All public methods shall start with tcfg_
Using a special prefix well help to ensure this module
has no collisions with existing methods when it is mixed in.
This is important for methods like 'set' and 'get' which are
very likely to collide with something in a base class.

The idea going forward will be to use TCFG_ prefix by convention
for
  * env var overrides
  * method names
  * config file names